+++ src/app/entrance/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import Button from "@/src/components/Button";
+import Button from "@/src/components/Common/Button";
 import { useRouter } from "next/navigation";
 
 function Entrance() {
